+++ src/main/java/store/controller/ConvenienceStore.java
@@ -1,4 +1,4 @@
-package store;
+package store.controller;
 
 import java.time.LocalDate;
 import java.util.List;
